# Make retrieving articles a little bit faster