Hi! Just a small performance update.

```ruby
random = rand(10_000..100_000)
array = Array(1..random);nil

Benchmark.ips do |benchmark|
  fetch_random = rand(1..random)
  benchmark.report('first(n)') { array.first(fetch_random);nil }
  benchmark.report('[0...n]')  { array[0...fetch_random];nil }
  benchmark.compare!
end
```

```sh
Warming up --------------------------------------
            first(n)   170.343k i/100ms
             [0...n]   139.688k i/100ms
Calculating -------------------------------------
            first(n)      5.212M (± 9.3%) i/s -     25.892M in   5.028039s
             [0...n]      2.878M (± 5.4%) i/s -     14.388M in   5.014773s

Comparison:
            first(n):  5212370.8 i/s
             [0...n]:  2878447.5 i/s - 1.81x  slower
```